Move identifier-case dry-run context into identifier-case module
Relocate identifier-case-context.js from src/plugin/src/reporting to src/plugin/src/identifier-case so the dry-run registry lives alongside the rest of the identifier-case pipeline rather than the reporting helpers.
