+++ src/plugin/src/identifier-case/identifier-case-context.js
