+++ src/plugin/src/reporting/identifier-case-report.js
@@ -10,7 +10,7 @@ import { setIdentifierCaseOption } from "../identifier-case/option-store.js";
 import { toTrimmedString } from "../../../shared/string-utils.js";
 import { asArray, toArray } from "../../../shared/array-utils.js";
 
-import { consumeIdentifierCaseDryRunContext } from "./identifier-case-context.js";
+import { consumeIdentifierCaseDryRunContext } from "../identifier-case/identifier-case-context.js";
 
 const REPORT_NAMESPACE = "gml-identifier-case";
 const LOG_VERSION = 1;
